Creating REST web application and client with PHP
Nowadays, web services are the ideal way to exchange data between a web application and another web or mobile application. REST, which stands for Representational State Transfer, is a good architectural choice for communication over the internet. It makes sharing data between different devices and apps easier and simpler compared to other methods like SOAP, CORBA, WSDL. REST is not dependent on any protocol, but almost every RESTful service uses HTTP as its underlying protocol.

In this tutorial, I will show you how to:

Create a REST web application that connects to a MySQL database to retrieve data.
Build a client to communicate with the web service.